% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_kastkaart.R
\name{bee_kastkaart}
\alias{bee_kastkaart}
\title{Creating a kastkaart after making split or catching swarm}
\usage{
bee_kastkaart(date, size, name, swarm, origin = NA, mated = NA, export = FALSE)
}
\arguments{
\item{date}{Date of creation (any format)}

\item{size}{Size of hive in frames}

\item{name}{Name of the queen/hive}

\item{swarm}{Yes or no}

\item{origin}{Origin of hive}

\item{mated}{Yes or no}

\item{export}{TRUE or FALSE}
}
\value{
a file that has the start of a kastkaart
}
\description{
Creating a kastkaart after making split or catching swarm
}
\examples{
bee_kastkaart("14/05/2021", 6, "Hennie", "No", "F1 Dikkie", "Yes")
bee_kastkaart("2022-02-16", 8, "Tantie", "Yes", export = TRUE)
}
