% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_kastkaart_update.R
\name{bee_kastkaart_update}
\alias{bee_kastkaart_update}
\title{Updating kastkaart after check up}
\usage{
bee_kastkaart_update(
  files,
  date,
  location,
  size = NA,
  brood = NA,
  swarmcells = NA,
  pollen = NA,
  honey = NA,
  extra1 = NA,
  extra2 = NA
)
}
\arguments{
\item{files}{File location from project location}

\item{date}{Date of checkup}

\item{location}{Location of the hive}

\item{size}{Size of the hive in frames}

\item{brood}{Amount of brood in frames}

\item{swarmcells}{Yes or no}

\item{pollen}{Amount of pollen in frames}

\item{honey}{Amount of honey in frames}

\item{extra1}{Extra note}

\item{extra2}{Extra note}
}
\value{
updated version of the kastkaart
}
\description{
Updating kastkaart after check up
}
\examples{
x <- bee_kastkaart("2022-02-16", 8, "Tantie", "Yes")
bee_kastkaart_update(x, "06/25/2021", "Kesteren", 12, honey = 3,
 extra1 = "Performed checkup to see honey stores")
}
