% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_mating_flight.R
\name{bee_mating_flight}
\alias{bee_mating_flight}
\title{calculate important dates for swarming.}
\usage{
bee_mating_flight(date, swarmcells = FALSE, larvae_size = NA)
}
\arguments{
\item{date}{Date a split was made, in dd/mm/yyyy format.}

\item{swarmcells}{If there were swarmcells present.}

\item{larvae_size}{Choose between 1 and 3: 1 = eggs, 2 = small larvae, 3 = big larvae. (only applicable for hives with swarmcells present)}
}
\value{
3 dates: 1 for when the queen is ripe, 1 for when the queen goes on mating flight, 1 for the day you can check the hive again.
}
\description{
calculate important dates for swarming.
}
\examples{
bee_mating_flight("01/06/2022", swarmcells = TRUE, larvae_size = 2)
bee_mating_flight("27/04/2022")
}
