% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_winterfood.R
\name{bee_winterfood}
\alias{bee_winterfood}
\title{Calculate the amount of Honey and pollen that need to be fed before winter}
\usage{
bee_winterfood(honey, pollen, bees)
}
\arguments{
\item{honey}{Number of frames}

\item{pollen}{Number of frames}

\item{bees}{Number of frames}
}
\value{
A calculated amount of honey and pollen
}
\description{
Calculate the amount of Honey and pollen that need to be fed before winter
}
\examples{
bee_winterfood(7, 2.5, 19)
}
